agents/flights_searcher/tools.py
# -*- coding: utf-8 -*-
# © Copyright IBM Corporation 2024, 2025
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

from langchain.tools import tool

@tool
def search_flights(
    departure_date: str,
    departure_city: str,
    arrival_city: str,
    departure_time: str,
    return_date: str,
    return_time: str,
):
    """
    Search for flights based on the provided parameters.
    """

    return "this is your flight number RK679"